backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.1.3
numpy==1.24.3
networkx==3.2.1
scikit-learn==1.3.2
python-louvain==0.16
python-multipart==0.0.6
pydantic==2.5.0
geopy==2.4.1
vaderSentiment==3.3.2
praw==7.7.1
python-dotenv==1.0.1

